+++ pkg/api/api_impl.go
@@ -27,65 +27,10 @@ import (
 )
 
 func validatePathTemplate(template string) []config.PathTemplate {
-	if template == "" {
-		return nil
+	if template != "" {
+		return config.ParsePathTemplate("./" + strings.ReplaceAll(template, "\\", "/"))
 	}
-	template = "./" + strings.ReplaceAll(template, "\\", "/")
-
-	parts := make([]config.PathTemplate, 0, 4)
-	search := 0
-
-	// Split by placeholders
-	for search < len(template) {
-		// Jump to the next "["
-		if found := strings.IndexByte(template[search:], '['); found == -1 {
-			break
-		} else {
-			search += found
-		}
-		head, tail := template[:search], template[search:]
-		placeholder := config.NoPlaceholder
-
-		// Check for a placeholder
-		switch {
-		case strings.HasPrefix(tail, "[dir]"):
-			placeholder = config.DirPlaceholder
-			search += len("[dir]")
-
-		case strings.HasPrefix(tail, "[name]"):
-			placeholder = config.NamePlaceholder
-			search += len("[name]")
-
-		case strings.HasPrefix(tail, "[hash]"):
-			placeholder = config.HashPlaceholder
-			search += len("[hash]")
-
-		default:
-			// Skip past the "[" so we don't find it again
-			search++
-			continue
-		}
-
-		// Add a part for everything up to and including this placeholder
-		parts = append(parts, config.PathTemplate{
-			Data:        head,
-			Placeholder: placeholder,
-		})
-
-		// Reset the search after this placeholder
-		template = template[search:]
-		search = 0
-	}
-
-	// Append any remaining data as a part without a placeholder
-	if search < len(template) {
-		parts = append(parts, config.PathTemplate{
-			Data:        template,
-			Placeholder: config.NoPlaceholder,
-		})
-	}
-
-	return parts
+	return nil
 }
 
 func validatePlatform(value Platform) config.Platform {